[LAB-88][Source Code][Overnight fee feature and postman collection]

+++ park_bo_mcs/src/main/java/labdsoft/park_bo_mcs/models/park/OvernightConfig.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class OvernightConfig {
     @Column(nullable = false)
-    private double overnightFee;
+    private Integer overnightFee;
 
     @Column(nullable = false)
     private boolean enabled;
